Removed unused bricks before falling back to harsher measures.
Running out of memory like this is pretty harsh.  Unlike hitting
*our* memory limit, the out of memory condition here implies the
system really is strained for memory.  Most memory allocators
start taking exponential time to find a free block when that
occurs.

For that reason, we really want to get rid of a *lot* of bricks.
Deleting bricks which are `next' to each other in the allocator's
lists relieves pressure on the OS' memory manager, which is having
trouble finding contiguous blocks large enough for our bricks.

Despite this new strategy, there's still a buglet that we'll
delete bricks which are still in use if it's absolutely necessary.
The alternative is to crash, or maybe abort the render entirely.
I kept this behavior so that we could potentially implement some
sort of more advanced scheme later, whereby the memory manager
informs a higher level, `I know you were using this brick, but I
had to delete it; reallocate it if you still need it.'.

With this commit, I can render the 2.9GB `magnitude' dataset on a
32-bit Mac build which has 2GB of RAM.  In rare circumstances, a
brick ends up missing at the highest LoD, seemingly due to the
deletion of a brick which is in use.

+++ Renderer/GPUMemMan/GPUMemMan.cpp
@@ -580,6 +580,34 @@ find_brick_with_usercount(ForwIter first, const ForwIter last,
   return first;
 }
 
+// Gets rid of *all* unused bricks.  Returns the number of bricks it deleted.
+size_t GPUMemMan::DeleteUnusedBricks() {
+  size_t removed = 0;
+  // This is a bit harsh.  erase() in the middle of a deque invalidates *all*
+  // iterators.  So we repeatedly search for unused bricks, until our search
+  // comes up empty.
+  // That said, erase() at the beginning or end of a deque only invalidates the
+  // pointed-to element.  You might consider an optimization where we delete
+  // all unused bricks from the beginning and end of the structure, before
+  // moving on to the exhaustive+expensive search.
+  bool found;
+  do {
+    found = false;
+    const Texture3DListIter& iter = find_brick_with_usercount(
+                                      m_vpTex3DList.begin(),
+                                      m_vpTex3DList.end(), 0
+                                    );
+    if(iter != m_vpTex3DList.end()) {
+      ++removed;
+      found = true;
+      Delete3DTexture(iter);
+    }
+  } while(!m_vpTex3DList.empty() && found);
+
+  MESSAGE("Got rid of %u unused bricks.", static_cast<unsigned int>(removed));
+  return removed;
+}
+
 // We don't have enough CPU memory to load something.  Get rid of a brick.
 void GPUMemMan::DeleteArbitraryBrick() {
   assert(!m_vpTex3DList.empty());
@@ -628,7 +656,17 @@ GLTexture3D* GPUMemMan::Get3DTexture(Dataset* pDataset,
         T_ERROR("This system does not have enough memory to render a brick.");
         return NULL;
       }
-      DeleteArbitraryBrick();
+      // Delete all bricks that aren't used.  If that ends up being nothing,
+      // then we're pretty screwed.  Stupidly choose a brick in that case.
+      if(0 == DeleteUnusedBricks()) {
+        WARNING("No bricks unused.  Falling back to "
+                "deleting bricks that ARE in use!");
+        // Delete up to 4 bricks.  We want to delete multiple bricks here
+        // because we'll temporarily need copies of the bricks in memory.
+        for(size_t i=0; i < 4 && !m_vpTex3DList.empty(); ++i) {
+          DeleteArbitraryBrick();
+        }
+      }
     }
   } while(!m_vpTex3DList.empty());
   // Can't happen, but to quiet compilers:
